Extracted recommendations implementation to separate module
