+++ pages/index.tsx
@@ -3,9 +3,9 @@ import { Layout } from '@components/common'
 import HomeAllProductsGrid from '@components/common/HomeAllProductsGrid'
 import type { GetStaticPropsContext, InferGetStaticPropsType } from 'next'
 import { useCustomerOrders } from "@framework/customer";
-import { OrderEdge, OrderLineItemEdge } from "@framework/schema";
 import React, { useEffect, useState } from "react";
 import useSearch from "@framework/product/use-search";
+import { collectOrderedBrands, getRecommendedProduct } from "@lib/recommendations";
 import _ from "lodash";
 
 export async function getStaticProps({
@@ -58,30 +58,12 @@ export default function Home({
   })
 
   useEffect(() => {
-    customerOrders?.orders.edges.map((order: OrderEdge) => { // Iterate through orders
-      order.node.lineItems.edges.map((item: OrderLineItemEdge) => { // Iterate through ordered items
-        if (item.node.variant?.product.vendor && orderedBrands.indexOf(item.node.variant.product.vendor) === -1) { // If vendor is known and doesn't exist in orderedBrands list yet
-          const vendor = item.node.variant.product.vendor
-          setOrderedBrands(previousOrderedBrands => [...previousOrderedBrands, vendor]) // Collect list of ordered brands
-        }
-      })
-    })
+    collectOrderedBrands(customerOrders, orderedBrands, setOrderedBrands);
   }, [customerOrders?.orders.edges, orderedBrands])
 
   useEffect(() => {
-    if (customerOrders?.orders.edges.length) {
-      const recommendedProduct = _.sample(recommendedProducts?.products) // Take random recommended product from brand
-      if (recommendedProduct) {
-        const productListCopy = [...productList]
-        const result = productListCopy.findIndex(product => { return product.id === recommendedProduct?.id }) // Check if recommended product already exists in list of products
-        if (result != -1) { productListCopy.splice(result, 1) } // If it exists remove it
-        recommendedProduct.description += '-recommended-' // Alter description of recommended product to mark it as recommended
-        if (!productListCopy.some(product => product.description?.includes('-recommended-'))) { // Check if no product has been recommended already
-          setProductList([recommendedProduct, ...productListCopy]) // Add recommended product to beginning of list
-        }
-      }
-    }
-  }, [customerOrders, productList, recommendedProducts?.products])
+    getRecommendedProduct(customerOrders, recommendedProducts, productList, setProductList);
+  }, [customerOrders, productList, recommendedProducts])
 
   return (
     <>
